adds create speech recognizer sample
